requirements.txt
annotated-types==0.6.0
asgiref==3.7.2
attrs==23.2.0
bandit==1.7.6
Django==5.0.1
django-bitfield==2.2.0
django-debug-toolbar==4.2.0
django-ninja==1.1.0
flake8==7.0.0
flake8-bandit==4.1.1
flake8-bugbear==24.1.17
flake8-builtins==2.2.0
flake8-comprehensions==3.14.0
flake8-print==5.0.0
gitdb==4.0.11
GitPython==3.1.41
gunicorn==21.2.0
isort==5.13.2
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
packaging==23.2
pbr==6.0.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pydantic==2.5.3
pydantic_core==2.14.6
pyflakes==3.2.0
Pygments==2.17.2
PyYAML==6.0.1
rich==13.7.0
six==1.16.0
smmap==5.0.1
sqlparse==0.4.4
stevedore==5.1.0
tqdm==4.66.1
typing_extensions==4.9.0
